pyproject.toml
[tool.poetry]
name = "time-travellm"
version = "0.1.0"
description = ""
authors = ["Brandi Seeley <brandi.n.seeley@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
datasets = ">=2.16.0,<3.0.0"
requests = "^2.31.0"
langchain-community = "^0.3.27"
langchain = "^0.3.27"
jq = "^1.10.0"
tiktoken = "^0.9.0"
langchain-openai = "^0.3.28"
langgraph = "^0.6.3"
psycopg2-binary = "^2.9.9"
openai = "^1.12.0"
ragas = "^0.2.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.10"
langchain-postgres = "^0.0.15"
flask = "^3.0.0"
langsmith = "^0.3.45"
rapidfuzz = "^3.13.0"
langchain-core = "^0.3.72"
rank-bm25 = "^0.2.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
